Renders an allelic discrimination viewer

Renders a reactive allelic discrimination plot that is suitable for assigning to an UI output slot.

renderADplot(inputId, label = NULL, adData,
        targetColumn = "target.dyeId", xAxisTarget = "FAM",
        yAxisTarget = "HEX", valueColumn = "endPointRFU",
        colorBy = "genotype", polar = FALSE, showLegend = FALSE,
        plotlyCode = NULL, cssFile = NULL, cssText = NULL,
        interactive = TRUE)

Arguments

inputId
The input slot that will be used to modify plot.

label
Display label for the control, or NULL for no label.

adData
Allelic descrimination data with RDML$AsTable() format.

targetColumn
Column name that contains axis splitting (dye or target).

xAxisTarget
X axis target (dye) name.

valueColumn
Column name that contains discrimination value (RFU or cq).

colorBy
Column name that contains color levels data.

polar
Enables polar coordinates.

showLegend
Show plot legend.

plotlyCode
Your quoted custom plotly code.

cssFile
Path to the css styles file.

cssText
CSS styles as text.

interactive
Should be this adPLot interactive or not.

See also

Examples

library(RDML)
rdml <- RDML$new(system.file("/extdata/test.rdml", package = "shinyMolBio"))

#> 
#> Loading experiment: All Wells
#> 	run: Amp Step 3_FAM
#> 
#> 	run: Amp Step 3_HEX
#> 
#> Combining Bio-Rad runs

tbl <- rdml$AsTable(endPointRFU = mean(tail(data$adp$fpoints$fluor, 5)))
tbl <- group_by(tbl, position)

#> Error in group_by(tbl, position): could not find function "group_by"

tbl <- mutate(tbl,
              genotype = paste(
                if (endPointRFU[1] > 400 && endPointRFU[2] > 400) "AG"
                else if (endPointRFU[1] > 400) "AA"
                else if (endPointRFU[2] > 400) "GG"
                else "NA"
              ))

#> Error in mutate(tbl, genotype = paste(if (endPointRFU[1] > 400 && endPointRFU[2] >     400) "AG" else if (endPointRFU[1] > 400) "AA" else if (endPointRFU[2] >     400) "GG" else "NA")): could not find function "mutate"

renderADplot("f", "aa", tbl, polar = TRUE, showLegend = TRUE)

#> Error in renderADplot("f", "aa", tbl, polar = TRUE, showLegend = TRUE): Assertion on 'colnames(adData)' failed: Must include the elements {position,target.dyeId,endPointRFU,genotype}.