Renders an amplification curves viewer

Renders a reactive PCR amplification plot that is suitable for assigning to an UI output slot.

renderAmpCurves(inputId, label = NULL, ampCurves, colorBy = NULL,
  linetypeBy = NULL, logScale = FALSE, showCq = FALSE, showLegend = FALSE,
  thBy = NULL, plotlyCode = NULL, cssFile = NULL, cssText = NULL,
  interactive = TRUE)

Arguments

inputId

The input slot that will be used to modify plot.

label

Display label for the control, or NULL for no label.

ampCurves

Amplification curves data with RDML$GetFData(long.table = TRUE) format.

colorBy

Column name that contains color levels data.

linetypeBy

Column name that contains linetype levels data.

logScale

Converts plot to log(RFU).

showCq

Shows Cq with dots (cq and quantFluor columns have to be provided!).

showLegend

Show plot legend.

thBy

Column name that separates threshold values (quantFluor column have to be provided!).

plotlyCode

Your quoted custom plotly code.

cssFile

Path to the css styles file.

cssText

CSS styles as text.

interactive

Should be this pcrPlate interactive or not.

See also

Author

Konstantin A. Blagodatskikh <k.blag@yandex.ru>

Examples

library(RDML)
rdml <- RDML$new(system.file("/extdata/test.rdml", package = "shinyMolBio"))
#> 
#> Loading experiment: All Wells
#> 	run: Amp Step 3_FAM
#> 
#> 	run: Amp Step 3_HEX
#> 
#> Combining Bio-Rad runs
curves <- renderAmpCurves("curves1", ampCurves = rdml$GetFData(long.table = TRUE))
curves