Run shinyMolBio Example Applications

Launch shinyMolBio example applications

runExample(example)

Arguments

example

The name of the example to run, or NA (the default) to list the available examples.

Author

Konstantin A. Blagodatskikh <k.blag@yandex.ru>

Examples

## Only run this example in interactive R sessions
if (interactive()) {
 # List all available examples
 runExample()

 # Run one of the examples
 runExample("pcrPlateInput")

 # Print the directory containing the code for all examples
 system.file("shiny-examples", package="shinyMolBio")
}
#> Error: Please run `runExample()` with a valid example app as an argument.
#> Valid examples are: 'GlobalGUI', 'pcrPlateInput'